requirements.txt
# Core Dependencies
PyQt6==6.7.0
PyQt6-sip==13.6.0

# Network Communication
# TCP/IP and UDP are built-in with Python's socket module

# Audio Processing
pyaudio==0.2.13
sounddevice==0.4.6
numpy==1.24.3

# Video Processing
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Screen Sharing
mss==9.0.1
Pillow==10.1.0

# Database
# sqlite3 is built-in with Python

# Encryption & Security
cryptography==41.0.7
PyNaCl==1.5.0

# System Notifications
plyer==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
